Move fused RIME to experimental package
