+++ africanus/experimental/rime/fused/__init__.py
@@ -1,0 +1,2 @@
+from africanus.experimental.rime.fused.specification import RimeSpecification  # noqa
+from africanus.experimental.rime.fused.core import rime  # noqa
