+++ africanus/experimental/rime/fused/common.py
