--- africanus/rime/fused/__init__.py
@@ -1,2 +1,0 @@
-from africanus.rime.fused.specification import RimeSpecification  # noqa
-from africanus.rime.fused.core import rime  # noqa
